Version 0.79: - Mimic the functionality of CMD_TOGGLE_ORIENTATION and CMD_RESET_ORIENTATION commands found in original firmware (allows for front roll configuration) - fix calculating motor current. Reported motor current scaled by /8 to allow maximum 2A to fit into 8 bit - allow setting maximum motor current (CMD_EXT_SET_MAX_MOTOR_CURRENT) that prevents voltage droop (and possible module brown-out reset) when motor is stalling. Also detect stalling when this limit is reached. - stall detection timeout (from hall sensors) can be configured with CMD_EXT_SET_STALL_DETECTION_TIMEOUT - new readable register (CMD_EXT_GET_TUNING_PARAMS) that contains slowdown factor, minimum slowdonw speed, motor stalling timeout, maximum current as well as recorded current when motor stalled previously - update README

+++ Core/Inc/main.h
@@ -71,8 +71,8 @@ void motor_stopped();
 void motor_stall_check();
 void pwm_start(uint32_t channel);
 void pwm_stop(uint32_t channel);
-uint32_t get_voltage();
-uint32_t get_motor_current();
+uint16_t get_voltage();
+uint16_t get_motor_current();
 
 /* USER CODE END EFP */
 
@@ -108,7 +108,6 @@ uint32_t get_motor_current();
 
 #define LOW1_PWM_CHANNEL TIM_CHANNEL_1
 #define LOW2_PWM_CHANNEL TIM_CHANNEL_4
-
 
 #define UART_DMA_BUF_SIZE        64      /* DMA circular buffer size in bytes */
 #define DMA_TIMEOUT_MS      10      /* DMA Timeout duration in msec */
